legacy/requirements.worker.txt
asyncpg==0.28.0
loguru==0.7.0
pika==1.3.2
pillow==10.0.0
python-dotenv==1.0.0
pydantic==2.7.1
sqlalchemy==2.0.20
torch==2.6.0
tqdm==4.66.1
transformers==4.32.1 